feat: rehearsal-config.json schema
