+++ packages/cli/package.json
@@ -18,6 +18,7 @@
   ],
   "scripts": {
     "build": "tsc -b",
+    "build:schema": "typescript-json-schema \"src/configs/rehearsal-config.ts\" IRehearsalConfig --out ./rehearsal-config-schema.json",
     "docs": "api-extractor run --typescript-compiler-folder ./node_modules/typescript && api-documenter markdown -i ./temp -o ./docs",
     "lint": "npm-run-all lint:*",
     "lint:eslint": "eslint --fix . --ext .ts",
@@ -53,6 +54,7 @@
   },
   "devDependencies": {
     "@types/which": "^2.0.1",
+    "typescript-json-schema": "^0.55.0",
     "vitest": "^0.28.4"
   },
   "packageManager": "pnpm@7.12.1",
