+++ packages/cli/rehearsal-config-schema.json
@@ -1,0 +1,117 @@
+{
+    "$schema": "http://json-schema.org/draft-07/schema#",
+    "properties": {
+        "migrate": {
+            "properties": {
+                "install": {
+                    "properties": {
+                        "dependencies": {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        },
+                        "devDependencies": {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        }
+                    },
+                    "type": "object"
+                },
+                "setup": {
+                    "properties": {
+                        "lint": {
+                            "properties": {
+                                "args": {
+                                    "items": {
+                                        "type": "string"
+                                    },
+                                    "type": "array"
+                                },
+                                "command": {
+                                    "type": "string"
+                                }
+                            },
+                            "type": "object"
+                        },
+                        "ts": {
+                            "properties": {
+                                "args": {
+                                    "items": {
+                                        "type": "string"
+                                    },
+                                    "type": "array"
+                                },
+                                "command": {
+                                    "type": "string"
+                                }
+                            },
+                            "type": "object"
+                        }
+                    },
+                    "type": "object"
+                }
+            },
+            "type": "object"
+        },
+        "upgrade": {
+            "properties": {
+                "install": {
+                    "properties": {
+                        "dependencies": {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        },
+                        "devDependencies": {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        }
+                    },
+                    "type": "object"
+                },
+                "setup": {
+                    "properties": {
+                        "lint": {
+                            "properties": {
+                                "args": {
+                                    "items": {
+                                        "type": "string"
+                                    },
+                                    "type": "array"
+                                },
+                                "command": {
+                                    "type": "string"
+                                }
+                            },
+                            "type": "object"
+                        },
+                        "ts": {
+                            "properties": {
+                                "args": {
+                                    "items": {
+                                        "type": "string"
+                                    },
+                                    "type": "array"
+                                },
+                                "command": {
+                                    "type": "string"
+                                }
+                            },
+                            "type": "object"
+                        }
+                    },
+                    "type": "object"
+                }
+            },
+            "type": "object"
+        }
+    },
+    "type": "object"
+}
+
